+++ lib/util/migration/migration_util.dart
@@ -8,6 +8,7 @@
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:autonomy_flutter/common/injector.dart';
 import 'package:autonomy_flutter/database/cloud_database.dart';
 import 'package:autonomy_flutter/database/entity/connection.dart';
 import 'package:autonomy_flutter/database/entity/persona.dart';
@@ -17,11 +18,13 @@ import 'package:autonomy_flutter/service/audit_service.dart';
 import 'package:autonomy_flutter/service/backup_service.dart';
 import 'package:autonomy_flutter/service/configuration_service.dart';
 import 'package:autonomy_flutter/service/iap_service.dart';
+import 'package:autonomy_flutter/util/constants.dart';
 import 'package:autonomy_flutter/util/device.dart';
 import 'package:autonomy_flutter/util/log.dart';
 import 'package:autonomy_flutter/util/migration/migration_data.dart';
 import 'package:autonomy_flutter/util/wallet_storage_ext.dart';
 import 'package:flutter/services.dart';
+import 'package:nft_collection/services/address_service.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class MigrationUtil {
@@ -47,8 +50,49 @@ class MigrationUtil {
     if ((await _cloudDB.personaDao.getDefaultPersonas()).isNotEmpty) {
       _iapService.restore();
     }
-
+    await _migrateViewOnlyAddresses();
     log.info("[migration] finished");
+  }
+
+  Future<void> _migrateViewOnlyAddresses() async {
+    if (_configurationService.getDidMigrateAddress()) return;
+
+    final manualConnections = await _cloudDB.connectionDao
+        .getConnectionsByType(ConnectionType.manuallyAddress.rawValue);
+    final needChecksumConnections = manualConnections
+        .where((element) => element.key != _tryChecksum(element.key))
+        .toList();
+
+    if (needChecksumConnections.isNotEmpty) {
+      final addressService = injector<AddressService>();
+      final checksumConnections = needChecksumConnections.map((e) {
+        final checksumAddress = _tryChecksum(e.key);
+        return e.copyWith(key: checksumAddress, accountNumber: checksumAddress);
+      }).toList();
+      final personaAddresses =
+          (await _cloudDB.addressDao.getAddressesByType(CryptoType.ETH.source))
+              .map((e) => e.address);
+      final connectionAddresses = manualConnections.map((e) => e.key);
+      checksumConnections.removeWhere((element) =>
+          personaAddresses.contains(element.key) ||
+          connectionAddresses.contains(element.key));
+      await _cloudDB.connectionDao.deleteConnections(needChecksumConnections);
+      await addressService
+          .deleteAddresses(needChecksumConnections.map((e) => e.key).toList());
+      await _cloudDB.connectionDao.insertConnections(checksumConnections);
+      await addressService
+          .addAddresses(checksumConnections.map((e) => e.key).toList());
+    }
+
+    _configurationService.setDidMigrateAddress(true);
+  }
+
+  String _tryChecksum(String address) {
+    try {
+      return address.getETHEip55Address();
+    } catch (_) {
+      return address;
+    }
   }
 
   Future<void> migrationFromKeychain() async {
